Got rid of a bunch of warnings Added some error checking for shaders and ShaderPrograms Lighting takes specular light into account Lighting math is more accurate (should look better with textured objects)

+++ engine/core/events/InputManager.cpp
@@ -35,7 +35,7 @@ void InputManager::PollEvents(SDL_Event& event)
 	Uint32 mouseButton = SDL_GetMouseState(&mousex, &mousey);
 
 	mouse.prevPosition = mouse.position;
-	mouse.position = Vec2(mousex, mousey);
+	mouse.position = Vec2(static_cast<float>(mousex), static_cast<float>(mousey));
 	
 }
